requirements.txt
aiogram
google-genai
aiosqlite
aiohttp